constants fixed

+++ app/src/main/java/com/yerayyas/superheromarvelinfo/network/ApiService.kt
@@ -14,14 +14,14 @@ interface ApiService {
         @Query("nameStartsWith") nameStartsWith: String,
         @Query("apikey") apiKey: String,
         @Query("hash") hash: String,
-        @Query("ts") ts: Int
+        @Query("ts") ts: Long
     ): Response<SuperheroDataResponse>
 
     @GET("/v1/public/characters")
     suspend fun getSuperheroes(
         @Query("apikey") apiKey: String,
         @Query("hash") hash: String,
-        @Query("ts") ts: Int
+        @Query("ts") ts: Long
     ): Response<SuperheroDataResponse>
 
     @GET("/v1/public/characters/{characterId}")
@@ -29,6 +29,6 @@ interface ApiService {
         @Path("characterId") characterId: Int,
         @Query("apikey") apiKey: String,
         @Query("hash") hash: String,
-        @Query("ts") ts: Int
+        @Query("ts") ts: Long
     ): Response<ImageDatasResult>
 }
